@@ -12,21 +12,21 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "babel-runtime": "^6.25.0",
+    "@babel/runtime": "^7.0.0-beta.48",
+    "@material-ui/core": "^1.1.0",
+    "babel-runtime": "^6.26.0",
     "chroma-js": "^1.3.6",
     "classnames": "^2.2.5",
     "datejs": "^1.0.0-rc3",
-    "history": "^3.0.0",
+    "history": "^4.7.2",
     "js-cookie": "^2.1.4",
     "lodash": "^4.17.10",
-    "material-ui": "^0.18.7",
     "moment": "^2.22.1",
     "moment-timezone": "^0.5.16",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0",
-    "react-mounter": "1.2.0",
-    "react-router-dom": "^4.0.0-beta.6",
-    "react-tap-event-plugin": "^2.0.1",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "suncalc": "^1.8.0",
     "xmldom": "^0.1.27",
     "xss": "^0.3.3"
